requirements.txt
python-decouple
Django>=4.2.0
djangorestframework>=3.14.0
django-filter>=23.2